Refactor: Moved Common file to a single folder
